Fix bugs in Product module.

+++ Ecommerce/src/Modules/Inventory/Ecommerce.Modules.Inventory.Application/Inventory/Features/Products/DeleteProduct/DeleteProductHandler.cs
@@ -34,7 +34,10 @@ namespace Ecommerce.Modules.Inventory.Application.Inventory.Features.Products.De
         {
             var imagesIds = await _imageRepository.GetAllImagesForProductAsync(request.ProductId, cancellationToken);
             await _productRepository.DeleteAsync(request.ProductId, cancellationToken);
-            await _blobStorageService.DeleteManyAsync(imagesIds.Select(i => i.ToString()), _containerName);
+            if(imagesIds.Any())
+            {
+                await _blobStorageService.DeleteManyAsync(imagesIds.Select(i => i.ToString()), _containerName);
+            }
             _logger.LogInformation("Product: {productId} was deleted by {@user}.",
                 request.ProductId, new { _contextService.Identity!.Username, _contextService.Identity!.Id });
         }
